Split BaseServiceImpl to better support TO Services
Services for TO models cannot do "save" and do not have repositories.
The "save()" method is trivial, just move it out of the interface and
reduce it to a protected method as it is only used inside services.

With this commit, most TO services can implement the BaseService
interface and enable their Controller counter part to just extend
BaseController.
