+++ src/main/java/org/dcsa/core/controller/BaseController.java
@@ -3,8 +3,9 @@ package org.dcsa.core.controller;
 import lombok.extern.slf4j.Slf4j;
 import org.dcsa.core.exception.*;
 import org.dcsa.core.service.BaseService;
-import org.springframework.data.r2dbc.BadSqlGrammarException;
 import org.springframework.http.HttpStatus;
+import org.springframework.r2dbc.BadSqlGrammarException;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ServerWebInputException;
 import reactor.core.publisher.Mono;
@@ -23,7 +24,8 @@ public abstract class BaseController<S extends BaseService<T, I>, T, I> {
         return getService().findById(id);
     }
 
-    @PostMapping()
+    @Transactional
+    @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public Mono<T> create(@Valid @RequestBody T t) {
         S s = getService();
@@ -33,6 +35,7 @@ public abstract class BaseController<S extends BaseService<T, I>, T, I> {
         return s.create(t);
     }
 
+    @Transactional
     @PutMapping("{id}")
     @ResponseStatus(HttpStatus.OK)
     public Mono<T> update(@PathVariable I id, @Valid @RequestBody T t) {
@@ -43,7 +46,8 @@ public abstract class BaseController<S extends BaseService<T, I>, T, I> {
         return s.update(t);
     }
 
-    @DeleteMapping()
+    @Transactional
+    @DeleteMapping
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public Mono<Void> delete(@RequestBody T t) {
         S s = getService();
@@ -53,6 +57,7 @@ public abstract class BaseController<S extends BaseService<T, I>, T, I> {
         return s.delete(t);
     }
 
+    @Transactional
     @DeleteMapping("{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public Mono<Void> deleteById(@PathVariable I id) {
