+++ src/main/java/org/dcsa/core/service/impl/ExtendedBaseServiceImpl.java
@@ -3,25 +3,9 @@ package org.dcsa.core.service.impl;
 import org.dcsa.core.extendedrequest.ExtendedRequest;
 import org.dcsa.core.repository.ExtendedRepository;
 import org.dcsa.core.service.ExtendedBaseService;
-import org.dcsa.core.util.ReflectUtility;
 import reactor.core.publisher.Flux;
 
-public abstract class ExtendedBaseServiceImpl<R extends ExtendedRepository<T, I>, T, I> extends BaseServiceImpl<R, T, I> implements ExtendedBaseService<T, I> {
-
-    private transient Class<T> modelClass;
-
-    @Override
-    public Class<T> getModelClass() {
-        if (modelClass == null) {
-            this.modelClass = ReflectUtility.getConcreteModelClassForService(this.getClass());
-        }
-        return modelClass;
-    }
-
-    @Override
-    public String getType() {
-        return getModelClass().getSimpleName();
-    }
+public abstract class ExtendedBaseServiceImpl<R extends ExtendedRepository<T, I>, T, I> extends BaseRepositoryBackedServiceImpl<R, T, I> implements ExtendedBaseService<T, I> {
 
     public I getIdOfEntity(T entity) {
         return getRepository().getIdOfEntity(entity);
